### 2018-02-11 #578 @whedon
[PRE REVIEW]: Dassie: a database of subject terms and hierarchies in the Library of Congress Subject Headings
**Submitting author:** @mhucka (<a href="http://orcid.org/0000-0001-9105-5960">Michael Hucka</a>)
**Repository:** <a href="https://github.com/casics/dassie" target ="_blank">https://github.com/casics/dassie</a>
**Version:** v1.0.2
