### 2018-02-13 #578 @mhucka
I'm not sure where to report this, so I'll try in this issue thread first: the PDF generated automatically by whedon is different from the preview I generated myself (c.f. https://github.com/openjournals/joss/issues/379#issuecomment-362362109) and I cannot figure out why. The difference is in extra whitespace showing up in what whedon produced. I've attached a marked-up PDF that I generated locally, with arrows pointing out where the differences are relative to the PDF produced by whedon. The consequence of these differences on the first page is that in whedon's PDF, the last reference on the 2nd page in the paper spills onto a 3rd page, which is admittedly a small thing, but it's ugly, and avoiding that kind of thing is exactly the reason I was trying to find a way to preview JOSS PDFs in the first place :-).

Is this issue recognizable to anyone here? Is there anything that might be causing the introducing of extra whitespace when whedon runs?

I suppose a possible cause is using a different Latex or Pandoc environment. Is it possible to see a log of whedon's run when it produce the output? That might help identify differences.

![paper-difference_page_1](https://user-images.githubusercontent.com/1450019/36168472-131281f6-10ae-11e8-8ced-a0281bf6cc7b.png)
