### 2018-02-14 #578 @arfon
Interesting. I'm not sure what's going on here exactly @mhucka. The version I produce locally (which ultimately is the version that becomes the final PDF) looks much closer to what you have:
 [10.21105.joss.00578.pdf](https://github.com/openjournals/joss-reviews/files/1722403/10.21105.joss.00578.pdf)

These are the Heroku buildpacks we use to generate the Whedon TeX environment: https://github.com/openjournals/whedon-api/blob/master/app.json#L52-L55 but I'm not sure why we're getting extra whitespace in the Whedon-generated version of the PDF.
