### 2018-02-16 #578 @mhucka
Thanks for looking into this @arfon – and sorry it's such a trivial issue. But if you're getting the expected output, and that's what ultimately will be used, then that's good enough for me. It might not be worth pursuing further at this time.

Thanks again!
