+++ README.md
@@ -345,3 +345,15 @@ Everyone is asked to read and respect the [code of conduct](CONDUCT.md) when par
 Funding for this and other CASICS work has come from the [National Science Foundation](https://nsf.gov) via grant NSF EAGER #1533792 (Principal Investigator: Michael Hucka).
 
 The photo of a dassie (a type of rock hyrax) at the top of this page came from [Wikipedia](https://commons.wikimedia.org/wiki/File:Procavia-capensis-Frontal.JPG). The author is Bjørn Christian Tørrissen, who made it available under the terms of the Creative Commons Attribution-Share Alike 3.0 Unported license.
+    
+<br>
+<div align="center">
+  <a href="https://www.nsf.gov">
+    <img width="105" height="105" src=".graphics/NSF.svg">
+  </a>
+  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+  <a href="https://www.caltech.edu">
+    <img width="100" height="100" src=".graphics/caltech-round.svg">
+  </a>
+</div>
