### 2018-03-27 #578 @arfon
> @arfon I'm not sure how to handle this, so I have to ask here: I neglected to include an NSF agency funding acknowledgment in the paper. How can I add it? Can I go ahead and modify the paper that's in the github repo, or is there a different procedure I should follow?

Please go ahead and add this to the paper. I've seen authors add an `#acknowledgements` section to their `paper.md` file in the past. This should compile OK.
