### 2018-05-08 #578 @jasonclark
@soodoku Would you have time to review the Dassie software? Thanks for considering.
