### 2018-05-09 #578 @jasonclark
@soodoku Sure, not a problem. Thanks for agreeing. I'll set up next steps and we'll go from there.
