### 2018-05-09 #578 @jasonclark
@soodoku The reviewer guidelines are available here: http://joss.theoj.org/about#reviewer_guidelines
