### 2018-07-01 casics/locterms#1 @soodoku
JOSS Review
Dear @mhucka:

three suggestions:

1.  When you write: "We converted a portion of the LCSH linked data graph into a database that makes explicit the “is-a” relationships between LCSH terms." ---> can you be more precise about what 'portion.'

2.  When you write: "Dassie’s database contents were generated by beginning with the RDF file for the LCSH linked data (Library of Congress 2016b)", it would be useful to link to the ingest script and workflow.

3.  The paper needs a good example or use case. It doesn't have to be long. But for now, I don't have a good idea of how it will be used. I understand what it does but not clear about the why. You know a bunch more. Adding an example or two or a concrete sentence of two about researcher needs would be useful.
