### 2018-07-01 casics/locterms#4 @soodoku
API
Dear @mhucka:

I think this can be improved:
"The dassie command line interface (in the bin subdirectory) can perform four operations: print descriptive information about one or more LCSH terms, trace the "is-a" hierarchy upward from a given LCSH term until it reaches terms that have no hypernyms, search for terms whose labels or notes contain a given string or regular expression, and print some summary statistics about the database."

Here's one way:

See https://github.com/themains/know_your_ip#using-kyip

Give a clear mapping of flag to functionality in an abstract way. For instance,

```
dassie -d ID to describe the term ID
dassie -f term to find a particular string. It ....
dassie - ? for printing summary stat. of what?
```
