### 2018-07-01 casics/locterms#5 @soodoku
Installation
On Windows, you can't run bash without the following:
https://www.howtogeek.com/249966/how-to-install-and-use-the-linux-bash-shell-on-windows-10/

So include directions for that.
