### 2018-07-01 #728 @soodoku
Please fix:

1. https://github.com/casics/dassie/issues/5
2. https://github.com/casics/dassie/issues/4
3. https://github.com/casics/dassie/issues/1
