### 2018-07-01 casics/locterms#5 @soodoku
ok. it doesn't seem it works on windows. just make that point explicit. or add dir. on how to make it work. when i run `dassie-server start` on ubuntu bash shell on windows, it conks with:

```
/dassie-server: line 13: $'\r': command not found
./dassie-server: line 16: $'\r': command not found
./dassie-server: line 19: $'\r': command not found
./dassie-server: line 27: $'\r': command not found
./dassie-server: line 30: $'\r': command not found
./dassie-server: line 32: $'\r': command not found
./dassie-server: line 45: $'\r': command not found
./dassie-server: line 47: $'\r': command not found
./dassie-server: line 48: syntax error near unexpected token `$'{\r''
'/dassie-server: line 48: `print_usage() {
```
