### 2018-07-01 casics/locterms#5 @soodoku
can't run this on linux either. tried the chmod etc. can you try your instructions on a clean machine and get back w/ screenshots of the process or a more step-by-step write-up.
