### 2018-09-10 #728 @mhucka
Unfortunately I won't have time to do anything on this until later at least in the year. I don't know what the requirements are for the paper submission in terms of timing; if it's necessary to withdraw, then that's okay since I really don't have an alternative at this time.
