### 2019-01-24 #728 @mhucka
My apologies for the long radio silence, and thank you for your considerable patience.

Although editing the manuscript for the content & clarification issues would be easy, solving the Windows portability issues are likely to take an unpredictable amount of time. I really wish I could do it; I don't actually have a strong dislike of Windows, so that's not the problem – the problem is simply lack of time. I am forced to admit even to myself that it's just not going to happen anytime soon. I am in time bankruptcy.

What is the procedure in a situation like this? Withdraw the paper? Or something else?
